feat: Column Registry — enforced schema catalog with AI/OLAP metadata
Mandatory governance-first column registry that serves as the single source
of truth for all Storable field definitions. Every field on every Storable
must be pre-approved in the registry — enforced at class-definition time.

New files:
- store/registry.py: ColumnDef (7 metadata categories), ColumnRegistry,
  RegistryError. Supports define(), get(), resolve() (prefix split),
  validate_class(), validate_instance(), introspection.
- store/columns/__init__.py: REGISTRY global instance
- store/columns/trading.py: symbol, price, quantity, side, pnl, etc.
- store/columns/finance.py: bid, ask, strike, volatility, notional, etc.
- store/columns/general.py: name, label, title, status, weight, etc.
- tests/test_registry.py: 56 tests covering enforcement, constraints,
  prefix resolution, introspection, and global registry integration.

Key features:
- Measures MUST have unit (enforced at define() time)
- All columns require role (dimension/measure/attribute) and description
- allowed_prefixes on base columns (e.g. trader_name IS-A name)
- AI metadata: synonyms, sample_values, semantic_type
- OLAP: role + aggregation + unit
- Governance: sensitivity, deprecated, tags
- Cross-layer: legend_type, dh_type_override hints

Modified files:
- store/base.py: __init_subclass__ enforcement hook
- bridge/type_mapping.py: registry-aware DH schema inference
- tests/test_bridge.py: RichItem.amount Decimal→float (registry canonical)
- tests/test_workflow.py: Counter.value int→float (registry canonical)
- tests/test_reactive.py: Config no longer Storable (plain dataclass)
- README.md: Column Registry section + updated project structure

All 402 tests pass (127+137+49+16+17+56).

+++ tests/test_workflow.py
@@ -85,7 +85,7 @@ def client(conn_info, _provision_users):
 @dataclass
 class Counter(Storable):
     name: str = ""
-    value: int = 0
+    value: float = 0
 
 
 @dataclass
